Backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-docx
pypdf
python-multipart